Use a Rust trait to perform mapping from C
The idea is to get to a point where instead of defining prefix-suffix
pairs, we can just implement the trait for the types we need. This is
the first step towards that, it keeps the prefix-suffix structure, but
they are all pretty much the same, with the exception of `&[u8]` which
needs to take a ref to the variable that it's converting from.